Find intercept correction for cell estimates

For a given geography g, there may be a value y_g which is the ground truth. You have C cells with estimates of y that may be biased. This function, proposed by Ghitza and Gelman and Ghitza, will find a intercept shift for all C cells to best fit the estimand y. It is the argmin of the sum of absolute values of the deviation.

posthoc_intercept(xi, ests, n, search = c(-5, 5))

Arguments

xi
The true target

ests
A vector of current estimate, by cell

n
The sample size of the estimates, by cell

search
The lower and upper endpoints of the interval to search

Source

FindDelta function at https://github.com/Catalist-LLC/unemployment/blob/master/unemployment_cps_mrp/helper_functions/GetYHat.R

See also

Author

Yair Ghitza

Examples

 biased_ests <- ccesMRPrun:::invlogit(rnorm(n = 100, mean = 1, sd = 1))
 sizes <- rbinom(n = 100, size = 100, prob = 0.1)
 tru <- 0.5

 posthoc_intercept(tru, biased_ests, sizes)

#> [1] -1.024221