Get MCMC draws of post-stratified estimate of demog x cd cells

Get MCMC draws of post-stratified estimate of demog x cd cells

poststrat_draws(
  model,
  poststrat_tgt,
  orig_data = NULL,
  question_lbl = attr(orig_data, "question"),
  area_var = "cd",
  count_var = "count",
  new_levels = FALSE
)

Arguments

model
stan model from fit_brms

poststrat_tgt
The poststratification target. It must contain the column count, which is treated as the number of trials in the binomial model.

orig_data
original survey data. This defaults to NULL but if supplied be used to (1) subset the poststratification, to areas only in the survey, and (2) label the question outcome.

question_lbl
A character string that indicates the outcome, e.g. a shorthand for the outcome variable. This is useful when you want to preserve the outcome or description of multiple models.

area_var
A character string for the variable name(s) for area to group and aggregate by. That is, the area of interest in MRP. Defaults to "cd"

count_var
A character string for the variable name for the population count in the poststrat_tgt dataframe. This will be renamed as if it is a trial count in the model. Defaults to "count".

new_levels
If there are new levels in the poststrat table that do not have coefficients in the survey data, should there be an extrapolation or assignment to 0s? The answer should almost always be No in MRP.

Value

A tidy dataset with qID x cd x iter number of rows, where qID is the number of questions (outcomes), cd is the number of geographies, and iter is the number of iterations estimated in the MCMC model. The demographic cells within a district are averaged across, and a MRP estimate is computed. It contains the columns

iter

The number of iterations

cd

The geography

qID

The question

p_mrp_est

The proportion of success, estimated by MRP.

Examples

class(fit_GA) # brms object

#> [1] "brmsfit"

head(acs_GA) # dataset

#> # A tibble: 6 x 8
#>    year cd    female educ       age          count clinton_vote clinton_vote_2p…
#>   <dbl> <chr>  <int> <fct>      <fct>        <dbl>        <dbl>            <dbl>
#> 1  2016 GA-01      0 HS or Less 18 to 24 ye…  1082        0.409            0.420
#> 2  2016 GA-01      0 HS or Less 18 to 24 ye…  6603        0.409            0.420
#> 3  2016 GA-01      0 HS or Less 18 to 24 ye… 16412        0.409            0.420
#> 4  2016 GA-01      0 Some Coll… 18 to 24 ye… 14732        0.409            0.420
#> 5  2016 GA-01      0 Some Coll… 18 to 24 ye…  1214        0.409            0.420
#> 6  2016 GA-01      0 4-Year     18 to 24 ye…  1810        0.409            0.420

drw_GA <- poststrat_draws(fit_GA, poststrat_tgt = acs_GA)