Fits and tidies MRP outputs in one step

Fits and tidies MRP outputs in one step

mrp_onestep(
  .formula,
  .data,
  poststrat_tgt,
  area_var = "cd",
  weight_var,
  add_on = NULL,
  ...
)

Arguments

.formula
Formula in binary y ~ (1|x1) + (1|x2) form.

.data
Individual-level dataset

poststrat_tgt
The poststratification target. It must contain the column count, which is treated as the number of trials in the binomial model.

area_var
Character for the variable(s) that corresponds to the area to aggregate to.

weight_var
Character for the variable that corresponds to weights.

add_on
Any area-level data to be merged with the output, for example validation data

...
Additional arguments to pass to the model fitting function, fit_brms()

Details

Combines fit_brms, poststrat_draws, and direct_est. See scatter_45 for options on visualization

Examples

if (FALSE) {
mrp_fit <- mrp_onestep(response ~ (1|educ) + (1|cd),
                       .data = cces_GA,
                       poststrat_tgt = acs_GA,
                       area_var = "cd",
                       weight_var = "weight_post",
                       add_on = elec_GA)

scatter_45(mrp_fit, clinton_vote, p_mrp_est,
           xlab = "Clinton Vote", ylab = "MRP Estimate")
}